reorganizing functions for readability
