+++ html/cart.html
@@ -1,3 +1,4 @@
+<!--TODO make this not ugly-->
 <!DOCTYPE html>
 <html>
 <head>
