+++ html/resultsPage.html
@@ -13,6 +13,7 @@
   <input id="autocomplete-submit" type="submit">
 </form>
 
+<!-- removing cart data for merge to master
 <iframe id= "cart-iframe" class="iframe" src="./cart.html" style="display:none;float:right;width:50%;height:800px"></iframe>
 <form id="cart-form" style="float:right">
   <div id="cart-button-div">
@@ -28,15 +29,19 @@
         } else if (iframeDisplay == "block"){
           iframe.style.display = "none"
         }
-      
       })
     </script>
   </div>
-</form>
+</form>-->
 
+<script type="text/javascript">
+  let saleData = [JSON.parse(sessionStorage.getItem("saleData"))]
+  let purchaseData = [JSON.parse(sessionStorage.getItem("purchaseData"))]
+</script>
 <script type="text/javascript" src="../src/autoComplete.js"></script> 
-<script type="text/javascript" src="../src/getCardData.js"></script>
+<script type="module" src="../src/getCardData.js"></script>
 <script type="text/javascript" src="../src/openCardTab.js"></script>
+<script type="module" src="../src/printCardData.js"></script>
 
 <h1 id="card-name" style="font-size: 1.5em;"><script type="text/javascript">document.write(decodeURIComponent(sessionStorage["submission"]))</script></h1>
 <p id="static-card-info" style="font-size: 1em;"></p>
